+++ docs/changelog.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 - `force_large` argument in the `pull` method, which can be used to handle large archive downloads
+- `add` method
 
 ### Deprecated
 - `get_remote_data_str` method. Use `listr` instead